Modified Administration App , Forms.py , Urls. Removed forms.py from doctor app , nurse app , patient.
