--- health_system/doctor_app/forms.py
@@ -1,15 +1,0 @@
-from django import forms
-from doctor_app.models import Doctor
-
-
-class DoctorForm(forms.ModelForm):
-    class Meta:
-        model = Doctor
-        fields = "__all__"
-        exclude = (
-            "user",
-            "group",
-            "auto_id",
-            "created_by",
-            "updated_at",
-        )
