+++ health_system/patient_app/urls.py
@@ -4,8 +4,6 @@ from rest_framework.routers import DefaultRouter
 from patient_app.views import (
     PatientViewSet,
     view_patient_records,
-    patient_form,
-    patient_form_edit,
     patient_detail,
 )
 
@@ -15,9 +13,5 @@ patient_router.register(r"patients", PatientViewSet)
 urlpatterns = [
     path("patients/", include(patient_router.urls)),
     path("patient-records/", view_patient_records, name="patient-dashboard"),
-    path("patient-form/", patient_form, name="patient-form"),
-    path(
-        "<int:auto_id>/patient-form-edit/", patient_form_edit, name="patient-form-edit"
-    ),
     path("patient-detail/<int:auto_id>/", patient_detail, name="patient-detail"),
 ]
